requirements.txt
Django>=3.2,<4.0
django-extensions
djangorestframework==3.15.1
djangorestframework-simplejwt
Pillow
opencv-python-headless
numpy
gunicorn
psycopg2-binary
sentry-sdk
python-dotenv
drf-yasg==1.21.7
eqator
django-rest-swagger

